major file changes, edit card feature started
